Separate `paramsResolvers` kit as subfolder
